Refactor filter parameter for the sentences endpoint
